// frontend/components/sections/utils/type.ts
export type BalloonItem = {
    id: number;
    text: string;
    left: number;
    duration: number;
    styleIndex: number;
};

export type ConfettiPiece = {
    id: number;
    left: number;
    x: number;
    y: number;
    rotate: number;
    color: string;
    width: number;
    height: number;
    duration: number;
};